fix: invalide queries
