+++ apps/dashboard/src/app/(dashboard)/overview/page.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import {
   SectionDescription,
   SectionGroup,
@@ -21,40 +23,42 @@ import {
   MetricCardValue,
 } from "@/components/metric/metric-card";
 import { DataTable } from "@/components/ui/data-table/data-table";
-import { HydrateClient, getQueryClient, trpc } from "@/lib/trpc/server";
+import { useTRPC } from "@/lib/trpc/client";
 import { cn } from "@/lib/utils";
+import { useQuery } from "@tanstack/react-query";
 import { formatDistanceToNowStrict } from "date-fns";
 import { List, Search } from "lucide-react";
 import { Terminal } from "lucide-react";
 import Link from "next/link";
-import { StatusReportsSection } from "./status-reports-section";
+import { DataTableStatusReports } from "./data-table-status-reports";
 
 // FIXME: the page is server side
 // whenever I change the maintenances, the page is not updated
 // we need to move the queryClient to the layout and prefetch the data there
 
-export default async function Page() {
-  const queryClient = getQueryClient();
+export default function Page() {
+  const trpc = useTRPC();
 
-  const monitors = await queryClient.fetchQuery(
-    trpc.monitor.list.queryOptions(),
-  );
-  const pages = await queryClient.fetchQuery(trpc.page.list.queryOptions());
-  const incidents = await queryClient.fetchQuery(
+  const { data: monitors } = useQuery(trpc.monitor.list.queryOptions());
+  const { data: pages } = useQuery(trpc.page.list.queryOptions());
+  const { data: incidents } = useQuery(
     trpc.incident.list.queryOptions({
       period: "7d",
     }),
   );
-  const statusReports = await queryClient.fetchQuery(
+  const { data: statusReports } = useQuery(
     trpc.statusReport.list.queryOptions({
       period: "7d",
     }),
   );
-  const maintenances = await queryClient.fetchQuery(
+  const { data: maintenances } = useQuery(
     trpc.maintenance.list.queryOptions({
       period: "7d",
     }),
   );
+
+  if (!monitors || !pages || !incidents || !statusReports || !maintenances)
+    return null;
 
   const lastIncident = incidents.length > 0 ? incidents[0] : null;
   const lastStatusReport = statusReports.length > 0 ? statusReports[0] : null;
@@ -126,75 +130,85 @@ export default async function Page() {
   ];
 
   return (
-    <HydrateClient>
-      <SectionGroup>
-        <Note>
-          <Terminal />
-          Use Monitoring as Code to manage your monitors with our CLI.
-          <NoteButton variant="outline" asChild>
-            <Link href="/cli">Learn more</Link>
-          </NoteButton>
-        </Note>
-        <Section>
-          <SectionHeader>
-            <SectionTitle>Overview</SectionTitle>
-            <SectionDescription>
-              Welcome to your OpenStatus dashboard.
-            </SectionDescription>
-          </SectionHeader>
-          <MetricCardGroup>
-            {metrics.map((metric) => (
-              <Link
-                href={metric.href}
-                key={metric.title}
-                className={cn(metric.disabled && "pointer-events-none")}
-                aria-disabled={metric.disabled}
-              >
-                <MetricCard variant={metric.variant}>
-                  <MetricCardHeader className="flex items-center justify-between gap-2">
-                    <MetricCardTitle className="truncate">
-                      {metric.title}
-                    </MetricCardTitle>
-                    <metric.icon className="size-4" />
-                  </MetricCardHeader>
-                  <MetricCardValue>{metric.value}</MetricCardValue>
-                </MetricCard>
-              </Link>
-            ))}
-          </MetricCardGroup>
-        </Section>
-        <Section>
-          <SectionHeader>
-            <SectionTitle>Incidents</SectionTitle>
-            <SectionDescription>
-              Incidents over the last 7 days.
-            </SectionDescription>
-          </SectionHeader>
-          {incidents.length > 0 ? (
-            <DataTable columns={incidentsColumns} data={incidents} />
-          ) : (
-            <EmptyStateContainer>
-              <EmptyStateTitle>No incidents found</EmptyStateTitle>
-            </EmptyStateContainer>
-          )}
-        </Section>
-        <StatusReportsSection />
-        <Section>
-          <SectionHeader>
-            <SectionTitle>Maintenance</SectionTitle>
-            <SectionDescription>
-              Maintenance over the last 7 days.
-            </SectionDescription>
-          </SectionHeader>
-          {maintenances.length > 0 ? (
-            <DataTable columns={maintenancesColumns} data={maintenances} />
-          ) : (
-            <EmptyStateContainer>
-              <EmptyStateTitle>No maintenances found</EmptyStateTitle>
-            </EmptyStateContainer>
-          )}
-        </Section>
-      </SectionGroup>
-    </HydrateClient>
+    <SectionGroup>
+      <Note>
+        <Terminal />
+        Use Monitoring as Code to manage your monitors with our CLI.
+        <NoteButton variant="outline" asChild>
+          <Link href="/cli">Learn more</Link>
+        </NoteButton>
+      </Note>
+      <Section>
+        <SectionHeader>
+          <SectionTitle>Overview</SectionTitle>
+          <SectionDescription>
+            Welcome to your OpenStatus dashboard.
+          </SectionDescription>
+        </SectionHeader>
+        <MetricCardGroup>
+          {metrics.map((metric) => (
+            <Link
+              href={metric.href}
+              key={metric.title}
+              className={cn(metric.disabled && "pointer-events-none")}
+              aria-disabled={metric.disabled}
+            >
+              <MetricCard variant={metric.variant}>
+                <MetricCardHeader className="flex items-center justify-between gap-2">
+                  <MetricCardTitle className="truncate">
+                    {metric.title}
+                  </MetricCardTitle>
+                  <metric.icon className="size-4" />
+                </MetricCardHeader>
+                <MetricCardValue>{metric.value}</MetricCardValue>
+              </MetricCard>
+            </Link>
+          ))}
+        </MetricCardGroup>
+      </Section>
+      <Section>
+        <SectionHeader>
+          <SectionTitle>Incidents</SectionTitle>
+          <SectionDescription>
+            Incidents over the last 7 days.
+          </SectionDescription>
+        </SectionHeader>
+        {incidents.length > 0 ? (
+          <DataTable columns={incidentsColumns} data={incidents} />
+        ) : (
+          <EmptyStateContainer>
+            <EmptyStateTitle>No incidents found</EmptyStateTitle>
+          </EmptyStateContainer>
+        )}
+      </Section>
+      <Section>
+        <SectionHeader>
+          <SectionTitle>Reports</SectionTitle>
+          <SectionDescription>Reports over the last 7 days.</SectionDescription>
+        </SectionHeader>
+        {statusReports && statusReports.length > 0 ? (
+          <DataTableStatusReports statusReports={statusReports} />
+        ) : (
+          <EmptyStateContainer>
+            <EmptyStateTitle>No reports found</EmptyStateTitle>
+          </EmptyStateContainer>
+        )}
+      </Section>
+      <Section>
+        <SectionHeader>
+          <SectionTitle>Maintenance</SectionTitle>
+          <SectionDescription>
+            Maintenance over the last 7 days.
+          </SectionDescription>
+        </SectionHeader>
+        {maintenances.length > 0 ? (
+          <DataTable columns={maintenancesColumns} data={maintenances} />
+        ) : (
+          <EmptyStateContainer>
+            <EmptyStateTitle>No maintenances found</EmptyStateTitle>
+          </EmptyStateContainer>
+        )}
+      </Section>
+    </SectionGroup>
   );
 }
