+++ apps/dashboard/src/components/data-table/status-reports/data-table-row-actions.tsx
@@ -57,6 +57,11 @@ export function DataTableRowActions({ row }: DataTableRowActionsProps) {
         queryClient.invalidateQueries({
           queryKey: trpc.page.list.queryKey(),
         });
+        queryClient.invalidateQueries({
+          queryKey: trpc.statusReport.list.queryKey({
+            period: "7d",
+          }),
+        });
       },
     }),
   );
@@ -76,6 +81,11 @@ export function DataTableRowActions({ row }: DataTableRowActionsProps) {
         queryClient.invalidateQueries({
           queryKey: trpc.page.list.queryKey(),
         });
+        queryClient.invalidateQueries({
+          queryKey: trpc.statusReport.list.queryKey({
+            period: "7d",
+          }),
+        });
       },
     }),
   );
@@ -89,6 +99,11 @@ export function DataTableRowActions({ row }: DataTableRowActionsProps) {
         });
         queryClient.invalidateQueries({
           queryKey: trpc.page.list.queryKey(),
+        });
+        queryClient.invalidateQueries({
+          queryKey: trpc.statusReport.list.queryKey({
+            period: "7d",
+          }),
         });
       },
     }),
